+++ importers/src/Service/VendorService/OverDrive/OverDriveMagazinesVendorService.php
@@ -6,40 +6,31 @@
 
 namespace App\Service\VendorService\OverDrive;
 
-use App\Exception\UnknownVendorServiceException;
-use App\Service\DataWell\SearchService;
+use App\Service\DataWell\DataWellClient;
+use App\Service\VendorService\AbstractDataWellVendorService;
 use App\Service\VendorService\OverDrive\Api\Client;
-use App\Service\VendorService\ProgressBarTrait;
-use App\Service\VendorService\VendorServiceImporterInterface;
-use App\Service\VendorService\VendorServiceTrait;
-use App\Utils\Message\VendorImportResultMessage;
-use App\Utils\Types\IdentifierType;
-use App\Utils\Types\VendorStatus;
 use Psr\Cache\InvalidArgumentException;
 
 /**
  * Class OverDriveMagazinesVendorService.
  */
-class OverDriveMagazinesVendorService implements VendorServiceImporterInterface
+class OverDriveMagazinesVendorService extends AbstractDataWellVendorService
 {
-    use ProgressBarTrait;
-    use VendorServiceTrait;
-
     protected const VENDOR_ID = 16;
-
-    private const VENDOR_SEARCH_TERM = 'facet.acSource="ereolen magazines"';
     private const VENDOR_MAGAZINE_URL_BASE = 'link.overdrive.com';
+
+    protected array $datawellQuery = ['facet.acSource="ereolen magazines"'];
 
     /**
      * OverDriveMagazinesVendorService constructor.
      *
-     * @param SearchService $searchService
+     * @param DataWellClient $datawell
      *   Datawell search service
      * @param Client $apiClient
      *   Api client for the OverDrive API
      */
     public function __construct(
-        private readonly SearchService $searchService,
+        protected readonly DataWellClient $datawell,
         private readonly Client $apiClient
     ) {
     }
@@ -47,79 +38,22 @@ class OverDriveMagazinesVendorService implements VendorServiceImporterInterface
     /**
      * {@inheritdoc}
      */
-    public function load(): VendorImportResultMessage
+    protected function extractData(array $jsonContent): array
     {
-        if (!$this->vendorCoreService->acquireLock($this->getVendorId(), $this->ignoreLock)) {
-            return VendorImportResultMessage::error(self::ERROR_RUNNING);
-        }
+        $pidArray = $this->datawell->extractData($jsonContent);
 
-        $this->loadConfig();
+        // Get the OverDrive APIs title urls from the results
+        $pidTitleUrlArray = array_map('self::getTitleUrlFromDatableIdentifiers', $pidArray);
+        $pidTitleUrlArray = array_filter($pidTitleUrlArray);
 
-        $status = new VendorStatus();
+        // Get the OverDrive APIs crossRefIds from the results
+        $pidTitleIdArray = array_map('self::getTitleIdFromUrl', $pidTitleUrlArray);
+        $pidTitleIdArray = array_filter($pidTitleIdArray);
 
-        $this->progressStart('Search data well for: "'.self::VENDOR_SEARCH_TERM.'"');
+        // Get the OverDrive cover urls
+        $pidCoverUrlArray = array_map('self::getCoverUrl', $pidTitleIdArray);
 
-        $offset = 1;
-        try {
-            do {
-                $this->progressMessage('Search data well for: "'.self::VENDOR_SEARCH_TERM.'" (Offset: '.$offset.')');
-
-                // Search the data well for material with acSource set to "ereolen magazines".
-                [$pidResultArray, $more, $offset] = $this->searchService->search(self::VENDOR_SEARCH_TERM, $offset);
-
-                // Get the OverDrive APIs title urls from the results
-                $pidTitleUrlArray = array_map('self::getTitleUrlFromDatableIdentifiers', $pidResultArray);
-                $pidTitleUrlArray = array_filter($pidTitleUrlArray);
-
-                // Get the OverDrive APIs crossRefIds from the results
-                $pidTitleIdArray = array_map('self::getTitleIdFromUrl', $pidTitleUrlArray);
-                $pidTitleIdArray = array_filter($pidTitleIdArray);
-
-                // Get the OverDrive cover urls
-                $pidCoverUrlArray = array_map('self::getCoverUrl', $pidTitleIdArray);
-                $pidCoverUrlArray = array_filter($pidCoverUrlArray);
-
-                // Remove null values
-                array_filter($pidCoverUrlArray);
-
-                $batchSize = count($pidCoverUrlArray);
-                $this->vendorCoreService->updateOrInsertMaterials($status, $pidCoverUrlArray, IdentifierType::PID, $this->getVendorId(), $this->withUpdatesDate, $this->withoutQueue, $batchSize);
-
-                $this->progressMessageFormatted($status);
-                $this->progressAdvance();
-
-                if ($this->limit && $status->records >= $this->limit) {
-                    $more = false;
-                }
-            } while ($more);
-
-            $this->vendorCoreService->releaseLock($this->getVendorId());
-
-            return VendorImportResultMessage::success($status);
-        } catch (\Exception $exception) {
-            return VendorImportResultMessage::error($exception->getMessage());
-        }
-    }
-
-    /**
-     * Set config from service from DB vendor object.
-     *
-     * @throws UnknownVendorServiceException
-     */
-    private function loadConfig(): void
-    {
-        $vendor = $this->vendorCoreService->getVendor($this->getVendorId());
-
-        $clientId = $vendor->getDataServerUser();
-        $clientSecret = $vendor->getDataServerPassword();
-        $libraryAccountEndpoint = $vendor->getDataServerURI();
-
-        if (!empty($clientId) && !empty($clientSecret) && !empty($libraryAccountEndpoint)) {
-            $this->apiClient->setLibraryAccountEndpoint($libraryAccountEndpoint);
-            $this->apiClient->setCredentials($clientId, $clientSecret);
-        } else {
-            throw new \InvalidArgumentException('Missing configuration');
-        }
+        return array_filter($pidCoverUrlArray);
     }
 
     /**