+++ importers/CHANGELOG.md
@@ -11,6 +11,19 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
+### Added
+- Abstract DatawellVendor.
+- BlockBuster vendor service.
+- ComicsPlus vendor service.
+
+### Changed
+- Deprecated "datawell" vendor (now "ComicsPlus" vendor).
+- TheMovieDatabaseVendor has been refactored to use the Abstract DatawellVendor.
+- OverDriveMagazinesVendorService has been refactored to use the Abstract DatawellVendor.
+- PressReaderVendorService has been refactored to use the Abstract DatawellVendor.
+- Consolidated all datawell search code to a DataWellClient;
+- Ordered services.yaml and changed "datawell.*" from parameters to direct injection.
+
 ## [3.1.0] - 2022-08-22
 
 ### Added